Move AwsS3V4Signer to auth
