+++ .changes/next-release/feature-AmazonS3-e4cb21c.json
@@ -1,0 +1,5 @@
+{
+    "category": "Amazon S3", 
+    "type": "feature", 
+    "description": "Move `AWSS3V4Signer` to auth module."
+}
