+++ auth/src/main/java/software/amazon/awssdk/auth/signer/internal/SignerConstant.java
@@ -13,11 +13,11 @@
  * permissions and limitations under the License.
  */
 
-package software.amazon.awssdk.auth.signer;
+package software.amazon.awssdk.auth.signer.internal;
 
-import software.amazon.awssdk.annotations.SdkProtectedApi;
+import software.amazon.awssdk.annotations.SdkInternalApi;
 
-@SdkProtectedApi
+@SdkInternalApi
 public final class SignerConstant {
 
     public static final String AWS4_TERMINATOR = "aws4_request";
