+++ codegen/src/main/java/software/amazon/awssdk/codegen/poet/builder/BaseClientBuilderClass.java
@@ -209,7 +209,7 @@ public class BaseClientBuilderClass implements ClassSpec {
 
     private CodeBlock s3SignerDefinitionMethodBody() {
         return CodeBlock.of("return $T.create();\n",
-                            ClassName.get("software.amazon.awssdk.services.s3", "AwsS3V4Signer"));
+                            ClassName.get("software.amazon.awssdk.auth.signer", "AwsS3V4Signer"));
     }
 
     @Override
