+++ services/s3/src/main/java/software/amazon/awssdk/services/s3/handlers/DisableDoubleUrlEncodingInterceptor.java
@@ -15,7 +15,7 @@
 
 package software.amazon.awssdk.services.s3.handlers;
 
-import software.amazon.awssdk.auth.signer.AwsSignerExecutionAttribute;
+import software.amazon.awssdk.auth.signer.internal.AwsSignerExecutionAttribute;
 import software.amazon.awssdk.core.interceptor.Context;
 import software.amazon.awssdk.core.interceptor.ExecutionAttributes;
 import software.amazon.awssdk.core.interceptor.ExecutionInterceptor;
